See description for changes(WARNING:AUTH&DATA CHANGED>SEE NEW TEMPLATES)
- Introduced deleteCommandMessages option in data.json. By default this is false but when true all your command messages are deleted when executing any command. This is toggable in data.json which therefore requires an update from the updated template.

- Various commands have been rewritten to be asyncronous using async/await functions: strawpoll, weather, oxr, all nsfw commands, dictionary define, anime, google, google images, steam, scp, tpid and tpname.

- google command has been fixed, previously knowledge graph results would always error and be completely unused.

- wikipedia command has been removed. it was a jumbled mess that truly already gets covered by google.

- various response and error messages have been changed.

- lots of performance tweaks such as not defining extra constants, defining methods within classes instead of outside, using argument parsing instead of functions and so on.

- games command has been completely rewritten from scratch. Instead of using mobygames it now uses IGDB (IndieGames DB) which despite the name actually has info on pretty much all games and has a great API to boot. This does require a new API key so please update your auth.json with the new template data.

- Tweaks to webhook mentions so they are less prone to error and more likely to let you stalk people

- Removal of the superagent and request packages as I now use snekfetch everywhere

- Adding of the igdb-api-node package for the IGDB command

+++ src/commands/emojis/sendemoji.js
@@ -17,6 +17,7 @@
 
 const Discord = require('discord.js'),
 	commando = require('discord.js-commando'),
+	data = require('../../data.json'),
 	path = require('path');
 
 module.exports = class sendEmojiCommand extends commando.Command {
@@ -49,6 +50,11 @@ module.exports = class sendEmojiCommand extends commando.Command {
 
 	run (msg, args) {
 		msg.channel.send(args.message, {'files': [new Discord.MessageAttachment(path.join(__dirname, `images/${args.emojiName.toLowerCase()}.png`), `${args.emojiName}Emoji.png`)]});
-		msg.delete();
+	
+		if (msg.deletable && data.deleteCommandMessages) {
+			msg.delete();
+		}
+	
+		return	msg.delete();
 	}
 };
